major refactoring of directory
