+++ sysopt/__init__.py
@@ -1,6 +1,17 @@
-"""sysopt - System Modelling and Optimisation Library."""
+"""sysopt - Component-base system modelling and optimisation.
 
-from sysopt.types import *
-from sysopt.block import *
+Build component based models, symbolically evaluate them, and translate the
+result into desired backend format.
+
+Packages:
+    modelling - Classes and methods for constructing component based models.
+    problems  - API for solving model simulation and optimisation problems.
+    symbolic  - Classes and methods for building symbolic expressions.
+    backends  - Interface shims for different algorithmic/numerical routines.
+
+"""
+
+from sysopt.env import version
+from sysopt.modelling import *
 from sysopt.symbolic import *
-from sysopt.solver import *
+from sysopt.problems import *
