+++ sysopt/blocks/__init__.py
@@ -1,4 +1,3 @@
 """Common Blocks and Block factories."""
 
 from sysopt.blocks.common import *
-from sysopt.blocks.builders import *
