+++ sysopt/env.py
@@ -9,3 +9,12 @@ Notes:
 """
 
 backend = 'casadi'
+
+
+"""SysOpt version info."""
+major = 0
+minor = 5
+patch = 0
+
+
+version = f"{major}.{minor}.{patch}"
