+++ sysopt/helpers.py
@@ -1,7 +1,10 @@
 """Common helper functions."""
+from typing import List, Iterable, Type, Iterator
 
 
-def flatten(the_list, depth=1):
+def flatten(the_list: List, depth: int = 1) -> List:
+    """Recursively flatten the list to the given depth."""
+
     result = []
     for item in the_list:
         if depth > 0 and isinstance(item, (list, tuple)):
@@ -12,11 +15,13 @@ def flatten(the_list, depth=1):
     return result
 
 
-def strip_nones(the_list):
+def strip_nones(the_list: List) -> List:
+    """Returns a copy of the list with all Nones removed."""
     return [l_i for l_i in the_list if l_i is not None]
 
 
-def filter_by_class(iterable, cls):
+def filter_by_class(iterable: Iterable, cls: Type) -> Iterator:
+    """Gets an iterator that returns each instance of the given class."""
     for item in iterable:
         if isinstance(item, cls):
             yield item
