+++ sysopt/problems/solver.py
@@ -11,9 +11,9 @@ from sysopt.symbolic import (
     ExpressionGraph, Variable, Parameter, get_time_variable,
     is_symbolic, ConstantFunction, GraphWrapper
 )
-from sysopt.solver.canonical_transform import flatten_system
-from sysopt.symbolic.problem_data import Quadratures, ConstrainedFunctional, FlattenedSystem
-from sysopt.block import Block, Composite
+from sysopt.problems.canonical_transform import flatten_system
+from sysopt.problems.problem_data import Quadratures, ConstrainedFunctional, FlattenedSystem
+from sysopt.modelling.block import Block, Composite
 
 DecisionVariable = NewType('DecisionVariable', Union[Variable, Parameter])
