--- sysopt/solver/__init__.py
@@ -1,4 +1,0 @@
-"""Interface to symbolic solvers."""
-
-from sysopt.solver.solver import *
-from sysopt.solver.canonical_transform import *
