--- sysopt/version.py
@@ -1,7 +1,0 @@
-"""SysOpt version info."""
-major = 0
-minor = 0
-patch = 3
-
-
-version = f"{major}.{minor}.{patch}"
