+++ tests/0_core/test_wiring_api.py
@@ -2,7 +2,6 @@ import pytest
 
 from sysopt.blocks.common import Gain, Oscillator, LowPassFilter
 from sysopt import Composite
-from sysopt.block import check_wiring_or_raise
 from sysopt import exceptions
 
 
@@ -15,8 +14,8 @@ def test_wiring_port_to_port_succeeds():
         (gain.outputs, composite.outputs),
         (osc.outputs, gain.inputs)
     ]
-    
-    check_wiring_or_raise(composite)
+    composite.check_wiring_or_raise()
+    composite.check_wiring_or_raise()
 
 
 def test_wiring_channel_to_channel_succeeds():
@@ -29,7 +28,7 @@ def test_wiring_channel_to_channel_succeeds():
         (osc.outputs[0], gain.inputs[0])
     ]
 
-    check_wiring_or_raise(composite)
+    composite.check_wiring_or_raise()
 
 
 def test_wiring_channels_by_name_succeeds():
@@ -43,7 +42,7 @@ def test_wiring_channels_by_name_succeeds():
         (gain.outputs['output0'], composite.outputs['output0']),
         (osc.outputs['signal'], gain.inputs['input0'])
     ]
-    check_wiring_or_raise(composite)
+    composite.check_wiring_or_raise()
 
 
 def test_forwarding_is_valid():
@@ -56,7 +55,7 @@ def test_forwarding_is_valid():
         (gain.outputs, fltr.inputs),
         (fltr.outputs, composite.outputs)
     ]
-    check_wiring_or_raise(composite)
+    composite.check_wiring_or_raise()
 
 
 def test_forwarding_by_name_is_valid():
@@ -71,7 +70,7 @@ def test_forwarding_by_name_is_valid():
         (gain.outputs, fltr.inputs),
         (fltr.outputs, composite.outputs['out'])
     ]
-    check_wiring_or_raise(composite)
+    composite.check_wiring_or_raise()
 
 
 def test_unconnected_internal_input_should_throw():
@@ -85,7 +84,7 @@ def test_unconnected_internal_input_should_throw():
     ]
 
     with pytest.raises(exceptions.UnconnectedInputError):
-        check_wiring_or_raise(composite)
+        composite.check_wiring_or_raise()
 
 
 def test_no_forwarded_outputs_should_throw():
@@ -99,7 +98,7 @@ def test_no_forwarded_outputs_should_throw():
     ]
 
     with pytest.raises(exceptions.InvalidComponentError):
-        check_wiring_or_raise(composite)
+        composite.check_wiring_or_raise()
 
 
 def test_dangling_input_should_throw():
@@ -113,7 +112,7 @@ def test_dangling_input_should_throw():
     ]
     composite.declare_inputs(['in'])
     with pytest.raises(exceptions.DanglingInputError):
-        check_wiring_or_raise(composite)
+        composite.check_wiring_or_raise()
 
 
 def test_wrong_size_should_throw():
