+++ tests/2_solver_interface/test_codesign_problem.py
@@ -1,10 +1,10 @@
-import numpy as np
 import sympy as sp
-from sysopt.types import *
-from sysopt.block import Block
-from sysopt.solver.solver import Problem, SolverContext, get_time_variable
-from sysopt.block import Composite
-from sysopt.blocks import FullStateOutput, ConstantSignal
+from sysopt.modelling.types import *
+from sysopt.modelling.block import Block
+from sysopt.problems.solver import SolverContext, get_time_variable
+from sysopt.modelling.block import Composite
+from sysopt.modelling.builders import FullStateOutput
+from sysopt.blocks import ConstantSignal
 from sysopt.symbolic import Variable, PiecewiseConstantSignal
