+++ tests/2_solver_interface/test_expression_graph.py
@@ -7,7 +7,7 @@ from sysopt.symbolic.symbols import get_time_variable
 import sysopt.backends.scalar_ops as scalar_ops
 
 from sysopt.blocks import Gain, Oscillator
-from sysopt.block import Composite
+from sysopt.modelling.block import Composite
 
 import numpy as np
