+++ tests/4_backends/test_casadi_codesign.py
@@ -1,12 +1,12 @@
 import casadi
-import numpy as np
 import sympy as sp
-from sysopt.types import *
-from sysopt.block import Block
-from sysopt.solver.solver import Problem, SolverContext, get_time_variable
-from sysopt.block import Composite
-from sysopt.blocks import FullStateOutput, ConstantSignal
-from sysopt.symbolic import Variable, PiecewiseConstantSignal
+from sysopt.modelling.types import *
+from sysopt.modelling.block import Block
+from sysopt.problems.solver import SolverContext
+from sysopt.modelling.block import Composite
+from sysopt.modelling.builders import FullStateOutput
+from sysopt.blocks import ConstantSignal
+from sysopt.symbolic import PiecewiseConstantSignal
 from sysopt.backends.casadi.codesign_problem import (
     CasadiCodesignProblemData, transcribe_problem, CodesignSolverOptions
 )
