+++ tests/8_codesign_test_problems/test_problem_2.py
@@ -7,10 +7,11 @@ Mechanical Design 141.1 (2019).
 
 """
 
-from sysopt.blocks import FullStateOutput, ConstantSignal
+from sysopt.modelling.builders import FullStateOutput
+from sysopt.blocks import ConstantSignal
 from sysopt import Metadata, Composite
 from sysopt.symbolic import Parameter, PiecewiseConstantSignal
-from sysopt.solver import SolverContext
+from sysopt.problems import SolverContext
 
 
 def dxdt(t, x, u, p):
