+++ tests/9_examples/test_driven_duffing.py
@@ -17,7 +17,7 @@ from sysopt import Composite
 from sysopt.blocks import Gain, Oscillator
 
 # Used to run simulation
-from sysopt.solver import SolverContext
+from sysopt.problems import SolverContext
 
 # @nb.text_cell
 r"""
@@ -183,7 +183,7 @@ The name can be index directly from the objects parameter attribute,
 or specified as a string.  
 
 Once the parameters are fully specified, one can construct a simulation/
-solver window (using the `SolverContext` context manager) and then
+problems window (using the `SolverContext` context manager) and then
 perform the integration with the `integrate` method.
 
 The output `x_t` is a function that provides cubic interpolation between
